+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 #Changelog
+- fix more issues with open_end slots
+
 
 ## v0.6.1 on t2s2-dev
 - allow empty string for enddate if it's a slot with open end